Use hooks for telemetry init and history watching
